+++ ExonautExtension/src/main/java/xyz/openexonaut/extension/exolib/ExoSuit.java
@@ -5,21 +5,21 @@ import com.fasterxml.jackson.databind.*;
 public class ExoSuit {
     public final String Name;
     public final ExoMod WeaponMod;
-    public final int Health;
-    public final int Regen_Speed;
-    public final int Regen_Delay;
-    public final int Timer_Boosts;
-    public final int Timer_Sp_Weps;
-    public final int CoolDown_Sp_Weps;
+    public final float Health;
+    public final float Regen_Speed;
+    public final float Regen_Delay;
+    public final float Timer_Boosts;
+    public final float Timer_Sp_Weps;
+    public final float CoolDown_Sp_Weps;
 
     public ExoSuit(JsonNode node, ExoMod[] allMods) {
         this.Name = node.get("Name").asText();
         this.WeaponMod = allMods[node.get("WeaponMod").asInt() - 1];
-        this.Health = node.get("Health").asInt();
-        this.Regen_Speed = node.get("Regen_Speed").asInt();
-        this.Regen_Delay = node.get("Regen_Delay").asInt();
-        this.Timer_Boosts = node.get("Timer_Boosts").asInt();
-        this.Timer_Sp_Weps = node.get("Timer_Sp_Weps").asInt();
-        this.CoolDown_Sp_Weps = node.get("CoolDown_Sp_Weps").asInt();
+        this.Health = (float) node.get("Health").asDouble();
+        this.Regen_Speed = (float) node.get("Regen_Speed").asDouble();
+        this.Regen_Delay = (float) node.get("Regen_Delay").asDouble();
+        this.Timer_Boosts = (float) node.get("Timer_Boosts").asDouble();
+        this.Timer_Sp_Weps = (float) node.get("Timer_Sp_Weps").asDouble();
+        this.CoolDown_Sp_Weps = (float) node.get("CoolDown_Sp_Weps").asDouble();
     }
 }
